backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
motor==3.6.0
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1
beautifulsoup4==4.12.3
openai==1.59.8
lxml==5.3.0
textstat==0.7.3
nltk==3.8.1
playwright==1.40.0
pillow==10.0.0